Restructure repo to match BRD layout
