+++ graphs/brd_graph.py
@@ -1,0 +1,9 @@
+"""LangGraph state machine for business requirement drafting."""
+
+
+class BRDGraph:
+    """Placeholder for graph implementation."""
+
+    def build(self) -> None:
+        """Construct the state machine."""
+        pass
